Add files via upload
I found out why the Time was an error because as I looked at it for a few minutes, I noticed that waitTimeStr = input("How many hours do you want to wait" is not closed because it suppose to be waitTimeStr = input("How many hours do you want to wait "). As with other errors and that is currentTimeInt = int(current_time_str) and
waitTimeInt = int(wait_time_str) because the variables are supposed to be named currentTimeStr and waitTimeStr. Especially the last two are wrong. finalTimeInt = currentTimeInt + waitTimeInt
print(finalTime_Int). The variable have to  be finalTimeInt = (currentTimeInt + waitTimeInt) % 24
print("The time after waiting will be:", finalTimeInt)